requirements.txt
python-dev 
libffi-dev 
libssl-dev 
libxml2-dev 
libxslt1-dev
Scrapy
pyyaml
pymongo
pybloom